@@ -1,18 +1,18 @@
 module github.com/corazawaf/coraza/v3/examples/http-server
 
-go 1.22.0
+go 1.23.0
 
 require github.com/corazawaf/coraza/v3 v3.2.1
 
 require (
 	github.com/corazawaf/libinjection-go v0.2.2 // indirect
-	github.com/magefile/mage v1.15.0 // indirect
-	github.com/petar-dambovaliev/aho-corasick v0.0.0-20240411101913-e07a1f0e8eb4 // indirect
+	github.com/magefile/mage v1.15.1-0.20250615140142-78acbaf2e3ae // indirect
+	github.com/petar-dambovaliev/aho-corasick v0.0.0-20250424160509-463d218d4745 // indirect
 	github.com/tidwall/gjson v1.18.0 // indirect
 	github.com/tidwall/match v1.1.1 // indirect
 	github.com/tidwall/pretty v1.2.1 // indirect
-	golang.org/x/net v0.30.0 // indirect
-	golang.org/x/sync v0.8.0 // indirect
-	golang.org/x/tools v0.22.0 // indirect
+	golang.org/x/net v0.43.0 // indirect
+	golang.org/x/sync v0.16.0 // indirect
+	golang.org/x/tools v0.35.0 // indirect
 	rsc.io/binaryregexp v0.2.0 // indirect
 )
