@@ -1,8 +1,7 @@
 // Copyright 2022 Juan Pablo Tosso and the OWASP Coraza contributors
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build !tinygo
-// +build !tinygo
+//go:build !tinygo && !windows && !plan9
 
 package auditlog
 
@@ -17,6 +16,9 @@ func init() {
 	})
 	RegisterWriter("https", func() plugintypes.AuditLogWriter {
 		return &httpsWriter{}
+	})
+	RegisterWriter("syslog", func() plugintypes.AuditLogWriter {
+		return NewSyslogWriter()
 	})
 
 	RegisterFormatter("json", &jsonFormatter{})
