@@ -32,7 +32,7 @@ func newPM(options plugintypes.OperatorOptions) (plugintypes.Operator, error) {
 		DFA:                  true,
 	})
 
-	m, _ := memoize.Do(data, func() (interface{}, error) { return builder.Build(dict), nil })
+	m, _ := memoize.Do(data, func() (any, error) { return builder.Build(dict), nil })
 	// TODO this operator is supposed to support snort data syntax: "@pm A|42|C|44|F"
 	return &pm{matcher: m.(ahocorasick.AhoCorasick)}, nil
 }
