@@ -81,8 +81,8 @@ func TestTransformationsAreCaseInsensitive(t *testing.T) {
 func unmarshalTests(json []byte) []Test {
 	var tests []Test
 	v := gjson.ParseBytes(json).Value()
-	for _, in := range v.([]interface{}) {
-		obj := in.(map[string]interface{})
+	for _, in := range v.([]any) {
+		obj := in.(map[string]any)
 		t := Test{}
 		if s, ok := obj["input"]; ok {
 			t.Input = s.(string)
