@@ -3,7 +3,6 @@
 
 // tinygo does not support net.http so this package is not needed for it
 //go:build !tinygo
-// +build !tinygo
 
 package http
 
@@ -19,12 +18,14 @@ import (
 // rwInterceptor intercepts the ResponseWriter, so it can track response size
 // and returned status code.
 type rwInterceptor struct {
-	w                  http.ResponseWriter
-	tx                 types.Transaction
-	statusCode         int
-	proto              string
-	isWriteHeaderFlush bool
-	wroteHeader        bool
+	w                             http.ResponseWriter
+	tx                            types.Transaction
+	statusCode                    int
+	proto                         string
+	isWriteHeaderFlush            bool
+	wroteHeader                   bool
+	wroteBufferedBodyToDownstream bool
+	allowFlushing                 bool
 }
 
 // WriteHeader records the status code to be sent right before the moment
@@ -51,6 +52,11 @@ func (i *rwInterceptor) WriteHeader(statusCode int) {
 	}
 
 	i.wroteHeader = true
+	if !i.tx.IsResponseBodyAccessible() || !i.tx.IsResponseBodyProcessable() {
+		// if the response body isn't accessible or processable we can already allow flushing
+		// we need to set this flag before the first call to Flush()
+		i.allowFlushing = true
+	}
 }
 
 // overrideWriteHeader overrides the recorded status code
@@ -77,7 +83,7 @@ func (i *rwInterceptor) cleanHeaders() {
 // interruption is triggered, this buffer is later used to analyse the body in
 // the response processor.
 // If the body isn't accessible or the mime type isn't processable, the response
-// body is being writen to the delegate response writer directly.
+// body is being written to the delegate response writer directly.
 func (i *rwInterceptor) Write(b []byte) (int, error) {
 	if i.tx.IsInterrupted() {
 		// if there is an interruption it must be from at least phase 4 and hence
@@ -95,7 +101,7 @@ func (i *rwInterceptor) Write(b []byte) (int, error) {
 		i.WriteHeader(http.StatusOK)
 	}
 
-	if i.tx.IsResponseBodyAccessible() && i.tx.IsResponseBodyProcessable() {
+	if i.tx.IsResponseBodyAccessible() && i.tx.IsResponseBodyProcessable() && !i.wroteBufferedBodyToDownstream {
 		// we only buffer the response body if we are going to access
 		// to it, otherwise we just send it to the response writer.
 		it, n, err := i.tx.WriteResponseBody(b)
@@ -111,7 +117,14 @@ func (i *rwInterceptor) Write(b []byte) (int, error) {
 			// See https://pkg.go.dev/io#Writer
 			return len(b), nil
 		}
-		return n, err
+		if err != nil || n == len(b) {
+			return n, err
+		}
+		if err := i.writeBufferedResponseBodyToDownstream(); err != nil {
+			return n, err
+		}
+		n2, err := i.w.Write(b[n:])
+		return n + n2, err
 	}
 
 	// flush the status code before writing
@@ -134,6 +147,41 @@ func (i *rwInterceptor) Flush() {
 	if !i.wroteHeader {
 		i.WriteHeader(http.StatusOK)
 	}
+
+	if i.allowFlushing {
+		if i.isWriteHeaderFlush {
+			// only propagate flush if the headers have been flushed already
+			if fl, ok := i.w.(http.Flusher); ok {
+				fl.Flush()
+			}
+		}
+
+	}
+}
+
+func (i *rwInterceptor) writeBufferedResponseBodyToDownstream() error {
+	if i.wroteBufferedBodyToDownstream {
+		return nil
+	}
+
+	// we release the buffer
+	reader, err := i.tx.ResponseBodyReader()
+	if err != nil {
+		i.overrideWriteHeader(http.StatusInternalServerError)
+		i.flushWriteHeader()
+		return fmt.Errorf("failed to release the response body reader: %v", err)
+	}
+
+	// this is the last opportunity we have to report the resolved status code
+	// as next step is write into the response writer (triggering a 200 in the
+	// response status code.)
+	i.flushWriteHeader()
+	if _, err := io.Copy(i.w, reader); err != nil {
+		return fmt.Errorf("failed to copy the response body: %v", err)
+	}
+
+	i.wroteBufferedBodyToDownstream = true
+	return nil
 }
 
 type responseWriter interface {
@@ -165,7 +213,7 @@ func wrap(w http.ResponseWriter, r *http.Request, tx types.Transaction) (
 			return nil
 		}
 
-		if tx.IsResponseBodyAccessible() && tx.IsResponseBodyProcessable() {
+		if tx.IsResponseBodyAccessible() && tx.IsResponseBodyProcessable() && !i.wroteBufferedBodyToDownstream {
 			if it, err := tx.ProcessResponseBody(); err != nil {
 				i.overrideWriteHeader(http.StatusInternalServerError)
 				i.flushWriteHeader()
@@ -178,23 +226,9 @@ func wrap(w http.ResponseWriter, r *http.Request, tx types.Transaction) (
 				i.flushWriteHeader()
 				return nil
 			}
-
-			// we release the buffer
-			reader, err := tx.ResponseBodyReader()
-			if err != nil {
-				i.overrideWriteHeader(http.StatusInternalServerError)
-				i.flushWriteHeader()
-				return fmt.Errorf("failed to release the response body reader: %v", err)
-			}
-
-			// this is the last opportunity we have to report the resolved status code
-			// as next step is write into the response writer (triggering a 200 in the
-			// response status code.)
-			i.flushWriteHeader()
-			if _, err := io.Copy(w, reader); err != nil {
-				return fmt.Errorf("failed to copy the response body: %v", err)
-			}
+			return i.writeBufferedResponseBodyToDownstream()
 		} else {
+			i.allowFlushing = true
 			i.flushWriteHeader()
 		}
 
