@@ -30,6 +30,8 @@ import (
 
 type nativeFormatter struct{}
 
+type auditLogWithErrMesg interface{ ErrorMessage() string }
+
 func (nativeFormatter) Format(al plugintypes.AuditLog) ([]byte, error) {
 	if len(al.Parts()) == 0 {
 		return nil, nil
@@ -100,11 +102,21 @@ func (nativeFormatter) Format(al plugintypes.AuditLog) ([]byte, error) {
 			// Producer: ModSecurity for Apache/2.9.1 (http://www.modsecurity.org/).
 			// Server: Apache
 			// Engine-Mode: "ENABLED"
+
+			// AuditLogTrailer is also expected to contain the error message generated by the rule, if any
+			for _, alEntry := range al.Messages() {
+				alWithErrMsg, ok := alEntry.(auditLogWithErrMesg)
+				if ok && alWithErrMsg.ErrorMessage() != "" {
+					res.WriteByte('\n')
+					res.WriteString(alWithErrMsg.ErrorMessage())
+				}
+			}
+
 			_, _ = fmt.Fprintf(&res, "\nStopwatch: %s\nResponse-Body-Transformed: %s\nProducer: %s\nServer: %s", "", "", "", "")
 		case types.AuditLogPartRulesMatched:
-			for _, r := range al.Messages() {
+			for _, alEntry := range al.Messages() {
 				res.WriteByte('\n')
-				res.WriteString(r.Data().Raw())
+				res.WriteString(alEntry.Data().Raw())
 			}
 		}
 		res.WriteByte('\n')
