@@ -5,6 +5,7 @@ package bodyprocessors
 
 import (
 	"encoding/xml"
+	"errors"
 	"io"
 	"strings"
 
@@ -38,7 +39,7 @@ func readXML(reader io.Reader) ([]string, []string, error) {
 	dec.Entity = xml.HTMLEntity
 	for {
 		token, err := dec.Token()
-		if err != nil && err != io.EOF {
+		if err != nil && err != io.EOF && !isUnexpectedEOFXMLSyntaxError(err) {
 			return nil, nil, err
 		}
 		if token == nil {
@@ -56,6 +57,11 @@ func readXML(reader io.Reader) ([]string, []string, error) {
 		}
 	}
 	return attrs, content, nil
+}
+
+func isUnexpectedEOFXMLSyntaxError(err error) bool {
+	var serr *xml.SyntaxError
+	return errors.As(err, &serr) && serr.Msg == "unexpected EOF"
 }
 
 var (
