@@ -4,7 +4,6 @@
 // Audit logs are currently disabled for tinygo
 
 //go:build !tinygo
-// +build !tinygo
 
 package testing
 
@@ -183,7 +182,7 @@ func TestAuditLogRelevantOnlyNoAuditlog(t *testing.T) {
 	}
 }
 
-func TestAuditLogOnNoLog(t *testing.T) {
+func TestAuditLogOnWithNoLog(t *testing.T) {
 	waf := corazawaf.NewWAF()
 	parser := seclang.NewParser(waf)
 	if err := parser.FromString(`
@@ -194,7 +193,7 @@ func TestAuditLogOnNoLog(t *testing.T) {
 		SecAuditLogParts ABCHIJKZ
 		SecAuditLogRelevantStatus ".*"
 		# auditlog tells that the transaction will have to log matches meant to be logged (not the ones with nolog)
-		SecRule ARGS "@unconditionalMatch" "id:1,phase:1,nolog,auditlog,msg:'nolog message'"
+		SecRule ARGS "@unconditionalMatch" "id:1,phase:1,nolog,msg:'nolog message'"
 	`); err != nil {
 		t.Fatal(err)
 	}
@@ -331,5 +330,107 @@ func TestAuditLogRequestBody(t *testing.T) {
 	}
 	if req.Body() != params {
 		t.Fatalf("Expected %s uri, got %s", params, req.Body())
+	}
+}
+
+// Arule expected to be logged (log and auditlog flags enabled) should
+// print the error message in the audit log as part of the H section.
+func TestAuditLogHFlag(t *testing.T) {
+	waf := corazawaf.NewWAF()
+	parser := seclang.NewParser(waf)
+	if err := parser.FromString(`
+		SecRuleEngine DetectionOnly
+		SecAuditEngine On
+		SecAuditLogFormat json
+		SecAuditLogType serial
+		SecAuditLogParts AHZ
+		SecAuditLogRelevantStatus ".*"
+		# An audit log should contain messages section on H flag included
+		SecRule ARGS "@unconditionalMatch" "id:1,phase:1,log,auditlog,msg:'expected rule message'"
+	`); err != nil {
+		t.Fatal(err)
+	}
+	// generate a random tmp file
+	file, err := os.Create(filepath.Join(t.TempDir(), "tmp.log"))
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer os.Remove(file.Name())
+	if err := parser.FromString(fmt.Sprintf("SecAuditLog %s", file.Name())); err != nil {
+		t.Fatal(err)
+	}
+	tx := waf.NewTransaction()
+	tx.AddGetRequestArgument("test", "test")
+	tx.ProcessRequestHeaders()
+	// now we read file
+	if _, err := file.Seek(0, 0); err != nil {
+		t.Error(err)
+	}
+	tx.ProcessLogging()
+	var al auditlog.Log
+	if err := json.NewDecoder(file).Decode(&al); err != nil {
+		t.Error(err)
+	}
+	if len(al.Messages()) != 1 {
+		t.Fatalf("Expected 1 message, got %d", len(al.Messages()))
+	}
+	type auditLogWithErrMesg interface{ ErrorMessage() string }
+	alWithErrMsg, ok := al.Messages()[0].(auditLogWithErrMesg)
+	if !ok {
+		t.Fatalf("Expected message to be of type auditLogWithErrMesg")
+	}
+	expected := "expected rule message"
+	if !strings.Contains(alWithErrMsg.ErrorMessage(), expected) {
+		t.Errorf("Expected audit log to contain %q, got %q", expected, alWithErrMsg.ErrorMessage())
+	}
+}
+
+func TestAuditLogWithKFlagWithoutHFlag(t *testing.T) {
+	waf := corazawaf.NewWAF()
+	parser := seclang.NewParser(waf)
+	if err := parser.FromString(`
+		SecRuleEngine DetectionOnly
+		SecAuditEngine On
+		SecAuditLogFormat json
+		SecAuditLogType serial
+		SecAuditLogParts ABCKZ
+		SecAuditLogRelevantStatus ".*"
+		# auditlog should not contain error logs without H flag included
+		SecRule ARGS "@unconditionalMatch" "id:1,phase:1,log,auditlog,msg:'unexpected logged message'"
+	`); err != nil {
+		t.Fatal(err)
+	}
+	// generate a random tmp file
+	file, err := os.Create(filepath.Join(t.TempDir(), "tmp.log"))
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer os.Remove(file.Name())
+	if err := parser.FromString(fmt.Sprintf("SecAuditLog %s", file.Name())); err != nil {
+		t.Fatal(err)
+	}
+	tx := waf.NewTransaction()
+	tx.AddGetRequestArgument("test", "test")
+	tx.ProcessRequestHeaders()
+	// now we read file
+	if _, err := file.Seek(0, 0); err != nil {
+		t.Error(err)
+	}
+	tx.ProcessLogging()
+	var al auditlog.Log
+	if err := json.NewDecoder(file).Decode(&al); err != nil {
+		t.Error(err)
+	}
+	if len(al.Messages()) != 1 {
+		t.Fatalf("Expected 1 message, got %d", len(al.Messages()))
+	}
+	type auditLogWithErrMesg interface{ ErrorMessage() string }
+	alWithErrMsg, ok := al.Messages()[0].(auditLogWithErrMesg)
+	if !ok {
+		t.Fatalf("Expected message to be of type auditLogWithErrMesg")
+	}
+	notExpected := "unexpected logged message"
+	if strings.Contains(alWithErrMsg.ErrorMessage(), notExpected) {
+		t.Errorf("Not expected audit log to contain %q, got %q", notExpected, alWithErrMsg.ErrorMessage())
 	}
 }
